requirements.txt
absl-py==0.9.0
alabaster==0.7.12
allennlp==0.9.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
blessings==1.7
blinker==1.4
blis==0.2.4
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
cachetools==3.1.1
certifi==2020.6.20
cffi==1.13.1
chardet==3.0.4
click==7.1.2
conllu==1.3.1
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
editdistance==0.5.3
en-core-web-sm==2.2.0
entrypoints==0.3
filelock==3.0.12
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gensim==3.8.0
gevent==1.4.0
gluonnlp==0.8.1
google-api-core==1.16.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gpustat==0.6.0
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jieba==0.42.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
murmurhash==1.0.0
mxnet==1.5.1.post0
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numpy==1.17.2
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
overrides==2.5
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
plac==0.9.6
pluggy==0.13.1
prefetch-generator==1.0.1
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.4
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.3.1
python-dateutil==2.8.0
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
rank-bm25==0.2
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
responses==0.10.7
rouge==1.0.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.12.0
smart-open==1.11.1
snowballstemmer==2.0.0
spacy==2.1.9
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
statsmodels==0.10.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorboardX==2.1
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
tokenizers==0.5.2
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
transformers==2.5.1
Unidecode==1.1.1
urllib3==1.25.6
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
word2number==1.1
zipp==0.6.0
